requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-environ==0.4.5
django-heroku==0.0.0
<<<<<<< HEAD:requirements.txt
django-redis==5.0.0
=======
>>>>>>> 09688a53ca7bcc66f2488b418584a49bdf94d4c6:requirements.txt
django-seed==0.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Faker==8.2.0
gunicorn==20.1.0
idna==2.10
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
whitenoise==5.2.0
